pyproject.toml
[tool.poetry]
name = "job-bot"
version = "0.1.0"
description = ""
authors = ["Georgypython <nov_georgy@inbox.ru>"]
readme = "README.md"
packages = [{include = "job_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
flake8 = "^6.1.0"
aiogram = "^3.1.1"
pymongo = "^4.6.0"
python-dotenv = "^1.0.0"
python-dateutil = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
